Create new App.js and move old App to App_example
